refactor(api): extract ModConfigService from mods route handlers (#351)
Move file system operations and business logic (parseEnvFile,
updateEnvFile, parseModList) into a dedicated ModConfigService,
following the established service layer pattern (ConfigService,
HostnameService). Route handlers now delegate to the service
instead of directly using readFileSync/writeFileSync.
